Time: 6 ms (30.50%), Space: 6.2 MB (30.50%) - LeetHub

+++ 1359-count-all-valid-pickup-and-delivery-options/1359-count-all-valid-pickup-and-delivery-options.cpp
@@ -3,7 +3,7 @@ public:
     int countOrders(int n) {
        vector<long>dp(501);
         dp[1]=1;dp[2]=6;
-        long long mod=1e9+7;
+        int mod=1e9+7;
         for(int i=3;i<=500;i++)
         {
             long long odd= 2*i-1;
